Improve error handling in dealing with files (e.g. uploading).
For now we only handle the error when the max file size exceeds its
limits, but it should be fairly straightforward to add additional error
codes if we care about them.
