+++ common/files.ts
@@ -1,0 +1,3 @@
+export enum FilesErrorCode {
+  MaxFileSizeExceeded = 'MaxFileSizeExceeded',
+}
