+++ server/app.ts
@@ -12,9 +12,9 @@ import { container } from 'tsyringe'
 import { DISCORD_WEBHOOK_URL_TOKEN } from './lib/discord/webhook-notifier'
 import isDev from './lib/env/is-dev'
 import { errorPayloadMiddleware } from './lib/errors/error-payload-middleware'
-import { addMiddleware as fileStoreMiddleware, setStore } from './lib/file-upload'
-import AwsStore from './lib/file-upload/aws'
-import LocalFileStore from './lib/file-upload/local-filesystem'
+import { addMiddleware as fileStoreMiddleware, setStore } from './lib/files'
+import AwsStore from './lib/files/aws'
+import LocalFileStore from './lib/files/local-filesystem'
 import logMiddleware from './lib/logging/log-middleware'
 import log from './lib/logging/logger'
 import { updateEmailTemplates } from './lib/mail/update-templates'
