+++ server/lib/bugs/bugs-api.ts
@@ -12,8 +12,8 @@ import {
 import { SbUserId } from '../../../common/users/sb-user-id'
 import transact from '../db/transaction'
 import { DiscordWebhookNotifier } from '../discord/webhook-notifier'
-import { deleteFile, getSignedUrl, writeFile } from '../file-upload'
-import { handleMultipartFiles } from '../file-upload/handle-multipart-files'
+import { deleteFile, getSignedUrl, writeFile } from '../files'
+import { handleMultipartFiles } from '../files/handle-multipart-files'
 import { httpApi } from '../http/http-api'
 import { httpBefore, httpGet, httpPost } from '../http/route-decorators'
 import { JobScheduler } from '../jobs/job-scheduler'
