+++ server/lib/chat/chat-api.ts
@@ -28,8 +28,8 @@ import { CHANNEL_BANNERS } from '../../../common/flags'
 import { SbUser } from '../../../common/users/sb-user'
 import { SbUserId } from '../../../common/users/sb-user-id'
 import { asHttpError } from '../errors/error-with-payload'
-import { handleMultipartFiles } from '../file-upload/handle-multipart-files'
-import { MAX_IMAGE_SIZE } from '../file-upload/images'
+import { handleMultipartFiles } from '../files/handle-multipart-files'
+import { MAX_IMAGE_SIZE } from '../files/images'
 import { httpApi, httpBeforeAll } from '../http/http-api'
 import { httpBefore, httpDelete, httpGet, httpPatch, httpPost } from '../http/route-decorators'
 import { checkAllPermissions } from '../permissions/check-permissions'
