+++ server/lib/chat/chat-service.ts
@@ -41,8 +41,8 @@ import { DbClient } from '../db'
 import { FOREIGN_KEY_VIOLATION, UNIQUE_VIOLATION } from '../db/pg-error-codes'
 import transact from '../db/transaction'
 import { CodedError } from '../errors/coded-error'
-import { writeFile } from '../file-upload'
-import { createImagePath, resizeImage } from '../file-upload/images'
+import { writeFile } from '../files'
+import { createImagePath, resizeImage } from '../files/images'
 import { ImageService } from '../images/image-service'
 import logger from '../logging/logger'
 import filterChatMessage from '../messaging/filter-chat-message'
