+++ server/lib/files/files-error.ts
@@ -1,0 +1,4 @@
+import { FilesErrorCode } from '../../../common/files'
+import { CodedError } from '../errors/coded-error'
+
+export class FilesError extends CodedError<FilesErrorCode> {}
