+++ server/lib/files/public-assets-config.test.ts
