+++ server/lib/files/public-assets-config.ts
