+++ server/lib/leagues/league-api.ts
@@ -30,9 +30,9 @@ import { UNIQUE_VIOLATION } from '../db/pg-error-codes'
 import transact from '../db/transaction'
 import { CodedError, makeErrorConverterMiddleware } from '../errors/coded-error'
 import { asHttpError } from '../errors/error-with-payload'
-import { writeFile } from '../file-upload'
-import { handleMultipartFiles } from '../file-upload/handle-multipart-files'
-import { MAX_IMAGE_SIZE, createImagePath, resizeImage } from '../file-upload/images'
+import { writeFile } from '../files'
+import { handleMultipartFiles } from '../files/handle-multipart-files'
+import { MAX_IMAGE_SIZE, createImagePath, resizeImage } from '../files/images'
 import { httpApi, httpBeforeAll } from '../http/http-api'
 import { httpBefore, httpGet, httpPatch, httpPost } from '../http/route-decorators'
 import { checkAllPermissions } from '../permissions/check-permissions'
