+++ server/lib/maps/map-operations.ts
@@ -2,7 +2,7 @@ import { writeFile as fsWriteFile } from 'fs/promises'
 import { withFile as withTmpFile } from 'tmp-promise'
 import { MapInfo } from '../../../common/maps'
 import { SbUserId } from '../../../common/users/sb-user-id'
-import { readFile } from '../file-upload'
+import { readFile } from '../files'
 import { MapParseResult, parseMap } from '../maps/store'
 import { getMapInfo, updateParseData } from './map-models'
 import { MapParseData } from './parse-data'
