+++ server/lib/maps/store.ts
@@ -6,7 +6,7 @@ import Queue from '../../../common/async/promise-queue'
 import { isTestRun } from '../../../common/is-test-run'
 import { MapExtension, MapVisibility } from '../../../common/maps'
 import { SbUserId } from '../../../common/users/sb-user-id'
-import { writeFile } from '../file-upload'
+import { writeFile } from '../files'
 import { addMap } from './map-models'
 import { MapParseData } from './parse-data'
 import { MAP_PARSER_VERSION } from './parser-version'
