+++ server/lib/security/csp.ts
@@ -2,7 +2,7 @@ import crypto from 'crypto'
 import Koa from 'koa'
 import { container } from 'tsyringe'
 import isDev from '../env/is-dev'
-import { FileStoreType, PublicAssetsConfig } from '../file-upload/public-assets-config'
+import { FileStoreType, PublicAssetsConfig } from '../files/public-assets-config'
 
 const CSP_NONCE_VALUE = Symbol('cspNonceValue')
