+++ server/routes.ts
@@ -11,8 +11,8 @@ import { ServerConfig } from '../common/server-config'
 import { ClientSessionInfo } from '../common/users/session'
 import './http-apis'
 import isDev from './lib/env/is-dev'
-import { getUrl, readFile } from './lib/file-upload'
-import { FileStoreType, PublicAssetsConfig } from './lib/file-upload/public-assets-config'
+import { getUrl, readFile } from './lib/files'
+import { FileStoreType, PublicAssetsConfig } from './lib/files/public-assets-config'
 import { applyApiRoutes, resolveAllHttpApis } from './lib/http/http-api'
 import logger from './lib/logging/logger'
 import { NewsService } from './lib/news/news-service'
